streambuf::pbump takes int, so cast.

+++ include/sir.hh
@@ -600,7 +600,7 @@ namespace sir
 
                     auto this_write = std::min(static_cast<std::streamsize>(left), count - written);
                     memcpy(pptr(), s + written, this_write);
-                    pbump(this_write);
+                    pbump(static_cast<int>(this_write));
                     written += this_write;
                 } while (written < count);
